backend/pyproject.toml
[project]
name = "todo-evolution-backend"
version = "0.1.0"
description = "Backend API for the Todo Evolution project - Phase II Full-Stack Web Application"
authors = [
    {name = "metachemist", email = "metachemist@example.com"}
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "alembic>=1.13.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "uuid>=1.30",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
    "psycopg2-binary>=2.9.10"
]
readme = "README.md"
requires-python = ">= 3.13"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
todo-api = "src.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]